+++ lib/core/src/display/object.rs
@@ -1,288 +1,17 @@
-pub mod display_object;
-
+pub mod transform;
 
 use crate::prelude::*;
 
-use crate::data::function::callback::*;
-use crate::data::opt_vec::OptVec;
 use crate::closure;
 use crate::data::dirty;
 use crate::data::dirty::traits::*;
+use crate::data::opt_vec::OptVec;
 use crate::system::web::group;
 
-use nalgebra::{Vector3, Vector4, Matrix4, Perspective3};
 use basegl_system_web::Logger;
-
-
-
-// TODO: Refactor and describe
-
-// === CloneRef ===
-
-pub trait CloneRef {
-    fn clone_ref(&self) -> Self;
-}
-
-
-
-
-// =================
-// === AxisOrder ===
-// =================
-
-/// Defines the order in which particular axis coordinates are processed. Used
-/// for example to define the rotation order in `DisplayObject`.
-#[derive(Clone,Debug)]
-pub enum AxisOrder { XYZ, XZY, YXZ, YZX, ZXY, ZYX }
-
-impl Default for AxisOrder {
-    fn default() -> Self { Self::XYZ }
-}
-
-
-
-// ======================
-// === TransformOrder ===
-// ======================
-
-/// Defines the order in which transformations (scale, rotate, translate) are
-/// applied to a particular object.
-#[derive(Clone,Debug)]
-pub enum TransformOrder {
-    ScaleRotateTranslate,
-    ScaleTranslateRotate,
-    RotateScaleTranslate,
-    RotateTranslateScale,
-    TranslateRotateScale,
-    TranslateScaleRotate
-}
-
-impl Default for TransformOrder {
-    fn default() -> Self { Self::ScaleRotateTranslate }
-}
-
-
-
-// =================
-// === Transform ===
-// =================
-
-/// Structure describing transform of an object, in particular its position, scale, and rotation.
-/// You can use methods like `matrix` to get a combined transformation matrix. Bear in mind that
-/// the matrix will always be recomputed from scratch. This structure does not contain any caching
-/// mechanisms.
-#[derive(Clone,Debug)]
-pub struct Transform {
-    pub position        : Vector3<f32>,
-    pub scale           : Vector3<f32>,
-    pub rotation        : Vector3<f32>,
-    pub transform_order : TransformOrder,
-    pub rotation_order  : AxisOrder,
-}
-
-impl Default for Transform {
-    fn default() -> Self {
-        let position        = Vector3::new(0.0,0.0,0.0);
-        let scale           = Vector3::new(1.0,1.0,1.0);
-        let rotation        = Vector3::new(0.0,0.0,0.0);
-        let transform_order = default();
-        let rotation_order  = default();
-        Self {position,scale,rotation,transform_order,rotation_order}
-    }
-}
-
-impl Transform {
-    /// Creates a new transformation object.
-    pub fn new() -> Self { default() }
-
-    /// Computes transformation matrix from the provided scale, rotation, and
-    /// translation components, based on the transformation and rotation orders.
-    pub fn matrix(&self) -> Matrix4<f32> {
-        let mut matrix = Matrix4::identity();
-        let matrix_ref = &mut matrix;
-        match self.transform_order {
-            TransformOrder::ScaleRotateTranslate => {
-                self.append_scale       (matrix_ref);
-                self.append_rotation    (matrix_ref);
-                self.append_translation (matrix_ref);
-            }
-            TransformOrder::ScaleTranslateRotate => {
-                self.append_scale       (matrix_ref);
-                self.append_translation (matrix_ref);
-                self.append_rotation    (matrix_ref);
-            }
-            TransformOrder::RotateScaleTranslate => {
-                self.append_rotation    (matrix_ref);
-                self.append_scale       (matrix_ref);
-                self.append_translation (matrix_ref);
-            }
-            TransformOrder::RotateTranslateScale => {
-                self.append_rotation    (matrix_ref);
-                self.append_translation (matrix_ref);
-                self.append_scale       (matrix_ref);
-            }
-            TransformOrder::TranslateRotateScale => {
-                self.append_translation (matrix_ref);
-                self.append_rotation    (matrix_ref);
-                self.append_scale       (matrix_ref);
-            }
-            TransformOrder::TranslateScaleRotate => {
-                self.append_translation (matrix_ref);
-                self.append_scale       (matrix_ref);
-                self.append_rotation    (matrix_ref);
-            }
-        }
-        matrix
-    }
-
-    /// Computes a rotation matrix from the provided rotation values based on
-    /// the rotation order.
-    pub fn rotation_matrix(&self) -> Matrix4<f32> {
-        let rx = Matrix4::from_scaled_axis(Vector3::x() * self.rotation.x);
-        let ry = Matrix4::from_scaled_axis(Vector3::y() * self.rotation.y);
-        let rz = Matrix4::from_scaled_axis(Vector3::z() * self.rotation.z);
-        match self.rotation_order {
-            AxisOrder::XYZ => rz * ry * rx,
-            AxisOrder::XZY => ry * rz * rx,
-            AxisOrder::YXZ => rz * rx * ry,
-            AxisOrder::YZX => rx * rz * ry,
-            AxisOrder::ZXY => ry * rx * rz,
-            AxisOrder::ZYX => rx * ry * rz,
-        }
-    }
-
-    fn append_translation(&self, m:&mut Matrix4<f32>) {
-        m.append_translation_mut(&self.position);
-    }
-
-    fn append_rotation(&self, m:&mut Matrix4<f32>) {
-        *m = self.rotation_matrix() * (*m);
-    }
-
-    fn append_scale(&self, m:&mut Matrix4<f32>) {
-        m.append_nonuniform_scaling_mut(&self.scale);
-    }
-}
-
-
-
-// =======================
-// === CachedTransform ===
-// =======================
-
-/// The same as `Transform` but with caching. It contains cached transformation matrix and dirty
-/// flags which are set after fields are modified. You can use the `update` function to recompute
-/// the matrix.
-#[derive(Derivative)]
-#[derivative(Clone(bound=""))]
-#[derivative(Debug(bound=""))]
-pub struct CachedTransform<OnChange> {
-    transform        : Transform,
-    transform_matrix : Matrix4<f32>,
-    origin           : Matrix4<f32>,
-    matrix           : Matrix4<f32>,
-    pub dirty        : dirty::SharedBool<OnChange>,
-    pub logger       : Logger,
-}
-
-impl<OnChange> CachedTransform<OnChange> {
-    pub fn new(logger:Logger, on_change:OnChange) -> Self {
-        let logger_dirty     = logger.sub("dirty");
-        let transform        = default();
-        let transform_matrix = Matrix4::identity();
-        let origin           = Matrix4::identity();
-        let matrix           = Matrix4::identity();
-        let dirty            = dirty::SharedBool::new(logger_dirty,on_change);
-        Self {transform,transform_matrix,origin,matrix,dirty,logger}
-    }
-
-    /// Update the transformation matrix and return information if the data was really updated.
-    pub fn update(&mut self, new_origin:Option<&Matrix4<f32>>) -> bool {
-        let is_dirty       = self.dirty.check_all();
-        let origin_changed = new_origin.is_some();
-        let changed        = is_dirty || origin_changed;
-        if changed {
-            group!(self.logger, "Update.", {
-                if is_dirty {
-                    self.transform_matrix = self.transform.matrix();
-                    self.dirty.unset_all();
-                }
-                new_origin.iter().for_each(|t| self.origin = **t);
-                self.matrix = self.origin * self.transform_matrix;
-            })
-        }
-        changed
-    }
-}
-
-
-// === Getters ===
-
-impl<OnChange> CachedTransform<OnChange> {
-    pub fn position(&self) -> &Vector3<f32> {
-        &self.transform.position
-    }
-
-    pub fn rotation(&self) -> &Vector3<f32> {
-        &self.transform.rotation
-    }
-
-    pub fn scale(&self) -> &Vector3<f32> {
-        &self.transform.scale
-    }
-
-    pub fn matrix(&self) -> &Matrix4<f32> {
-        &self.matrix
-    }
-
-    pub fn global_position(&self) -> Vector3<f32> {
-        (self.matrix * Vector4::new(0.0,0.0,0.0,1.0)).xyz()
-    }
-}
-
-
-// === Setters ===
-
-impl<OnChange:Callback0> CachedTransform<OnChange> {
-    pub fn position_mut(&mut self) -> &mut Vector3<f32> {
-        self.dirty.set();
-        &mut self.transform.position
-    }
-
-    pub fn rotation_mut(&mut self) -> &mut Vector3<f32> {
-        self.dirty.set();
-        &mut self.transform.rotation
-    }
-
-    pub fn scale_mut(&mut self) -> &mut Vector3<f32> {
-        self.dirty.set();
-        &mut self.transform.scale
-    }
-
-    pub fn set_position(&mut self, t:Vector3<f32>) {
-        *self.position_mut() = t;
-    }
-
-    pub fn set_rotation(&mut self, t:Vector3<f32>) {
-        *self.rotation_mut() = t;
-    }
-
-    pub fn set_scale(&mut self, t:Vector3<f32>) {
-        *self.scale_mut() = t;
-    }
-
-    pub fn mod_position<F:FnOnce(&mut Vector3<f32>)>(&mut self, f:F) {
-        f(self.position_mut())
-    }
-
-    pub fn mod_rotation<F:FnOnce(&mut Vector3<f32>)>(&mut self, f:F) {
-        f(self.rotation_mut())
-    }
-
-    pub fn mod_scale<F:FnOnce(&mut Vector3<f32>)>(&mut self, f:F) {
-        f(self.scale_mut())
-    }
-}
+use nalgebra::Vector3;
+use nalgebra::Matrix4;
+use transform::CachedTransform;
 
 
 
@@ -311,13 +40,15 @@ impl ParentBind {
 // === HierarchicalObjectDescription ===
 // =====================================
 
-///
+/// Hierarchical description of objects. Each object contains binding to its parent and to its
+/// children. This is the most underlying structure in the display object hierarchy.
 #[derive(Clone,Debug)]
 pub struct HierarchicalObjectData {
     pub parent_bind : Option<ParentBind>,
     pub children    : OptVec<DisplayObjectData>,
     pub logger      : Logger,
 }
+
 
 // === Public API ===
 
@@ -328,6 +59,7 @@ impl HierarchicalObjectData {
         Self {parent_bind,children,logger}
     }
 }
+
 
 // === Private API ===
 
@@ -351,6 +83,7 @@ impl HierarchicalObjectData {
     }
 }
 
+
 // === Getters ===
 
 impl DisplayObjectDataMut {
@@ -358,6 +91,7 @@ impl DisplayObjectDataMut {
         self.parent_bind.as_ref().map(|ref t| &t.parent)
     }
 }
+
 
 
 // ======================================
@@ -433,7 +167,7 @@ impl DisplayObjectDataMut {
                 self.logger.info("Self origin did not change.");
                 if self.child_dirty.check_all() {
                     group!(self.logger, "Updating dirty children.", {
-                        self.child_dirty.borrow().data.iter().for_each(|ix| {
+                        self.child_dirty.take().iter().for_each(|ix| {
                             self.children[*ix].update_with(origin,false)
                         });
                     })
@@ -445,6 +179,7 @@ impl DisplayObjectDataMut {
         if let Some(f) = &self.on_updated { f(self) }
     }
 }
+
 
 // === Private API ===
 
@@ -470,7 +205,7 @@ impl DisplayObjectDataMut {
 
     fn set_parent_bind(&mut self, bind:ParentBind) {
         self.logger.info("Adding new parent bind.");
-        let dirty     = bind.parent.rc.borrow().child_dirty.clone_rc();
+        let dirty     = bind.parent.rc.borrow().child_dirty.clone_ref();
         let on_change = fn_on_change(dirty, bind.index);
         self.transform.dirty.set_callback(Some(on_change.clone()));
         self.child_dirty.set_callback(Some(on_change));
@@ -566,6 +301,7 @@ impl DisplayObjectDataMut {
 pub struct DisplayObjectData {
     rc: Rc<RefCell<DisplayObjectDataMut>>,
 }
+
 
 // === Public API ==
 
@@ -756,6 +492,7 @@ impl CloneRef for DisplayObjectData {
 }
 
 
+
 // =====================
 // === DisplayObject ===
 // =====================
@@ -769,16 +506,14 @@ pub trait DisplayObject: Into<DisplayObjectData> {
 impl<T:Into<DisplayObjectData>> DisplayObject for T {}
 
 
-pub trait DisplayObjectOps where for<'t> &'t Self:DisplayObject {
-    fn add_child<T:DisplayObject>(&self, child:T) {
-        self.display_object_description().add_child(child)
-    }
-}
+//pub trait DisplayObjectOps where for<'t> &'t Self:DisplayObject {
+//    fn add_child<T:DisplayObject>(&self, child:T) {
+//        self.display_object_description().add_child(child)
+//    }
+//}
+//
+//impl<T> DisplayObjectOps for T where for<'t> &'t Self:DisplayObject {}
 
-impl<T> DisplayObjectOps for T where for<'t> &'t Self:DisplayObject {}
-
-
-// ==========================================================
 
 
 
@@ -791,21 +526,22 @@ mod tests {
     use super::*;
     use std::f32::consts::{PI};
 
-
     #[test]
     fn hierarchy_test() {
         let obj1 = DisplayObjectData::new(Logger::new("obj1"));
         let obj2 = DisplayObjectData::new(Logger::new("obj2"));
         let obj3 = DisplayObjectData::new(Logger::new("obj3"));
+        obj1.add_child(&obj2);
+        assert_eq!(obj2.index(),Some(0));
 
         obj1.add_child(&obj2);
-        assert_eq!(obj2.index(), Some(0));
-        obj1.add_child(&obj2);
-        assert_eq!(obj2.index(), Some(0));
+        assert_eq!(obj2.index(),Some(0));
+
         obj1.add_child(&obj3);
-        assert_eq!(obj3.index(), Some(1));
+        assert_eq!(obj3.index(),Some(1));
+
         obj1.remove_child(&obj3);
-        assert_eq!(obj3.index(), None);
+        assert_eq!(obj3.index(),None);
     }
 
     #[test]
@@ -813,13 +549,13 @@ mod tests {
         let obj1 = DisplayObjectData::new(Logger::new("obj1"));
         let obj2 = DisplayObjectData::new(Logger::new("obj2"));
         let obj3 = DisplayObjectData::new(Logger::new("obj3"));
-
         assert_eq!(obj1.position()        , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj2.position()        , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj3.position()        , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj1.global_position() , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(0.0,0.0,0.0));
+
         obj1.mod_position(|t| t.x += 7.0);
         obj1.add_child(&obj2);
         obj2.add_child(&obj3);
@@ -829,6 +565,7 @@ mod tests {
         assert_eq!(obj1.global_position() , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(0.0,0.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(0.0,0.0,0.0));
+
         obj1.update();
         assert_eq!(obj1.position()        , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj2.position()        , Vector3::new(0.0,0.0,0.0));
@@ -836,30 +573,37 @@ mod tests {
         assert_eq!(obj1.global_position() , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(7.0,0.0,0.0));
+
         obj2.mod_position(|t| t.y += 5.0);
         obj1.update();
         assert_eq!(obj1.global_position() , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(7.0,5.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(7.0,5.0,0.0));
+
         obj3.mod_position(|t| t.x += 1.0);
         obj1.update();
         assert_eq!(obj1.global_position() , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(7.0,5.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(8.0,5.0,0.0));
+
         obj2.mod_rotation(|t| t.z += PI/2.0);
         obj1.update();
         assert_eq!(obj1.global_position() , Vector3::new(7.0,0.0,0.0));
         assert_eq!(obj2.global_position() , Vector3::new(7.0,5.0,0.0));
         assert_eq!(obj3.global_position() , Vector3::new(7.0,6.0,0.0));
+
         obj1.add_child(&obj3);
         obj1.update();
         assert_eq!(obj3.global_position() , Vector3::new(8.0,0.0,0.0));
+
         obj1.remove_child(&obj3);
         obj3.update();
         assert_eq!(obj3.global_position() , Vector3::new(1.0,0.0,0.0));
+
         obj2.add_child(&obj3);
         obj1.update();
         assert_eq!(obj3.global_position() , Vector3::new(7.0,6.0,0.0));
+
         obj1.remove_child(&obj3);
         obj1.update();
         obj2.update();
