+++ lib/core/src/display/symbol/geometry/primitive/mesh/buffer/item.rs
@@ -1,9 +1,7 @@
-use crate::prelude::*;
-
 use crate::display::render::webgl::Context;
 use nalgebra::*;
 use web_sys::WebGlUniformLocation;
-use nalgebra::storage::Owned;
+
 
 
 // =============
